KL-143/rename: move directory stores in src
FeedPage 하위에 있던 stores 디렉토리를 src 하위로 변경하였습니다.
각 컴포넌트 내 useFeedStore의 import 경로를 변경하였습니다.
