+++ src/pages/FeedPage/components/BasicFilter/CategorySection.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { Checkbox } from 'antd'
 import PropTypes from 'prop-types'
 // import useKy from '../../../../hooks/useKy'
-import useFeedStore from '../../stores/useFeedStore'
+import useFeedStore from '../../../../stores/useFeedStore'
 import SubCategoryContainer from './SubCategoryContainer'
 import {
   SectionContainer,
