+++ src/pages/FeedPage/components/SelectedField/CityField.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { CloseOutlined } from '@ant-design/icons'
-import useFeedStore from '../../stores/useFeedStore'
+import useFeedStore from '../../../../stores/useFeedStore'
 import {
   BlueFieldTag,
   WhiteFieldTag,
